Reorganise token handlers
