+++ apps/frontend/src/App.tsx
@@ -13,6 +13,8 @@
 import React, { useEffect } from 'react';
 import AppRouter from '@/router/AppRouter';
 import { AuthProvider, AuthProviderProps } from 'react-oidc-context';
+import { WebStorageStateStore } from 'oidc-client-ts';
+import i18n from '@/i18n';
 import eduApi from '@/api/eduApi';
 import BBBFrame from '@/pages/ConferencePage/BBBIFrame';
 import EmbeddedIframes from '@/components/framing/EmbeddedIframes';
@@ -21,12 +23,11 @@ import useLmnApiStore from '@/store/useLmnApiStore';
 import lmnApi from '@/api/lmnApi';
 import useUserStore from '@/store/UserStore/UserStore';
 import Toaster from '@/components/ui/Toaster';
-import { WebStorageStateStore } from 'oidc-client-ts';
 import { HTTP_HEADERS } from '@libs/common/types/http-methods';
 import EDU_API_ROOT from '@libs/common/constants/eduApiRoot';
-import i18n from '@/i18n';
 import VDIFrame from './pages/DesktopDeployment/VDIFrame';
 import CommunityLicenseDialog from './pages/UserSettings/Info/CommunityLicenseDialog';
+import GlobalHooksWrapper from './components/GlobalHooksWrapper';
 
 const App = () => {
   const { eduApiToken } = useUserStore();
@@ -58,13 +59,15 @@ const App = () => {
 
   return (
     <AuthProvider {...oidcConfig}>
-      <BBBFrame />
-      <VDIFrame />
-      <AppRouter />
-      <EmbeddedIframes />
-      <NativeFrames />
-      <CommunityLicenseDialog />
-      <Toaster />
+      <GlobalHooksWrapper>
+        <BBBFrame />
+        <VDIFrame />
+        <AppRouter />
+        <EmbeddedIframes />
+        <NativeFrames />
+        <CommunityLicenseDialog />
+        <Toaster />
+      </GlobalHooksWrapper>
     </AuthProvider>
   );
 };
